backend/user-service/requirements-dev.txt
# Development Requirements for User Service
# Install with: pip install -r requirements-dev.txt

# Include production requirements
-r requirements.txt

# Development Tools
ipython==8.21.0                    # Enhanced Python shell
django-extensions==3.2.3           # Django development utilities
django-debug-toolbar==4.3.0        # Debug toolbar for development

# Testing & Quality Assurance
pytest==8.0.2                      # Testing framework
pytest-django==4.8.0               # Django testing utilities
pytest-cov==4.1.0                  # Coverage reporting
pytest-mock==3.12.0                # Mocking utilities
factory-boy==3.3.0                 # Test data generation
faker==22.6.0                      # Fake data generation
coverage==7.4.1                     # Code coverage

# Code Quality & Formatting
black==24.2.0                      # Code formatting
flake8==7.0.0                      # Linting
isort==5.13.2                      # Import sorting
pre-commit==3.6.2                  # Git hooks for code quality
mypy==1.8.0                        # Type checking

# Documentation
sphinx==7.2.6                      # Documentation generator
sphinx-rtd-theme==2.0.0            # Read the Docs theme

# Performance & Profiling
django-silk==5.1.0                 # Request profiling
memory-profiler==0.61.0            # Memory usage profiling

# Security
bandit==1.7.5                      # Security linting
safety==2.3.5                      # Dependency vulnerability scanning

# Monitoring & Debugging
sentry-sdk==2.7.0                  # Error tracking
structlog==24.1.0                  # Structured logging
